Move RepositoryCollectionCapabilityException from data to data-validation module
