+++ molgenis-data-validation/src/main/java/org/molgenis/data/validation/RepositoryCollectionCapabilityException.java
@@ -1,9 +1,11 @@
-package org.molgenis.data;
+package org.molgenis.data.validation;
+
+import org.molgenis.data.RepositoryCapability;
 
 import static java.util.Objects.requireNonNull;
 
 @SuppressWarnings("squid:MaximumInheritanceDepth")
-public class RepositoryCollectionCapabilityException extends UnknownDataException
+public class RepositoryCollectionCapabilityException extends DataIntegrityViolationException
 {
 	private static final String ERROR_CODE = "D13";
